backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
alembic>=1.13.2
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
python-multipart>=0.0.9
aiohttp>=3.9.5
redis>=5.0.0
hiredis>=2.3.2
celery>=5.4.0
googlemaps>=4.10.0
openrouteservice>=2.3.0
python-dateutil>=2.9.0
pyodbc>=5.1.0
ortools>=9.9.0
numpy>=1.24.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4